// src/app/model/gestionCobranza/CortesServicioRequest.ts
export class CortesServicioRequest {
  id_corte: number = 0;
  id_tipo: number = 0;
  id_cliente: number = 0;
  motivo_corte: string = '';
  url_fotocorte: string = '';
  id_fotocorte: string = '';
  fecha_corte: string = '';
  tipo_corte: number = 0;
  responsable_corte: string = '';
  estado_corte: number = 0;
  fecha_reconexion: string = '';
  observaciones: string = '';
  fechareg: string = '';
  creador: string = '';
  nombre_completo: string = '';
  nomestado_corte: string = '';
  tipo_servicio: string = '';
  nrodocident:string='';
  op:number=0;
  id_reapertura:number=0;
}